More detailed exception types
1. Use custom exception types instead of the generic ones
2. Mesh Vertices/Indices size in bytes should be uint64_t in case we ever load larger models > 4 Mb

+++ projects/Common/src/Model.cpp
@@ -5,8 +5,8 @@ module Model;
 
 namespace gg
 {
-	uint32_t Mesh::VerticesSizeBytes() const { return static_cast<uint32_t>(vertices.size()) * sizeof(Vertex); }
-	uint32_t Mesh::IndicesSizeBytes() const { return static_cast<uint32_t>(indices.size()) * sizeof(uint32_t); }
+	uint64_t Mesh::VerticesSizeBytes() const { return static_cast<uint64_t>(vertices.size()) * sizeof(Vertex); }
+	uint64_t Mesh::IndicesSizeBytes() const { return static_cast<uint64_t>(indices.size()) * sizeof(uint32_t); }
 	uint32_t Mesh::GetVertexCount() const { return static_cast<uint32_t>(vertices.size()); }
 	uint32_t Mesh::GetIndexCount() const { return static_cast<uint32_t>(indices.size()); }
